NewModes: Add BatteryEquioment and Vehicle CHargingEquipment
The VDV review comments put forward Vehicle charging elements. to manage the charging of electric vehicles.    Adding a charging  plan for vehciles needs further modelling so should be left to a future CR.  However it is straightforward to add a BatteryEquipment  and a ChargingEquipmentProfile based on the VDV values, and incorporating ISOstandard plug types.