Extended-DatedVehicleJourney+Replace-PR368
Initial PR368 from Kristian Syversen (Kirstian has moved and the PR was not editable ... so it was easier to create a new one).
This is the exact copy of PR368, but targeting 'next' and not master branch and with deprecated DriverRef (DatedJourney are expected to be referenced by driver's DUTY, not the way arround).
This is conforming the decision from metting #3 of the NeTEx revision 2022/2023.