Retype all order elements to xsd:positiveInteger and fix the associated examples
This is a breaking change (with respect to the documentation) which mixes xsd:positiveInteger with xsd:integer.

The main difference is: positiveInteger starts at 1.

Please carefully consider the changes that have a mention on "Relative order". No examples exists with a negative relative order, but the design might allow them?

Fix #475