Review of the CHARGING PLAN
Review of the CHARGING PLAN's NeTEx implementation
References to VEHICLE and VEHICLE TYPE has been updated to VEHICLEs and VEHICLE TYPEs (multiple possible). Furthermore, this is a small extension to Transmodel (or a more direct access, avoiding the need to walk through the object hierarchy). It's fine an useful to me, but probably need double check.